correction long range

+++ ft_rangelong_base.c
@@ -6,15 +6,15 @@
 /*   By: tmanet <tmanet@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/10 10:25:32 by tmanet            #+#    #+#             */
-/*   Updated: 2015/12/10 10:44:43 by tmanet           ###   ########.fr       */
+/*   Updated: 2017/02/06 14:04:44 by tmanet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_rangelong_base(unsigned long nbr, unsigned int base)
+unsigned long	ft_rangelong_base(unsigned long nbr, unsigned int base)
 {
-	int range;
+	unsigned long	range;
 
 	range = 1;
 	while (nbr >= (base))
